git-svn-id: svn://192.168.1.8/ontom/trunk@3272 85aebdb3-f732-af43-9dbd-bcef24ea9265

+++ ontom/blueprint.c
@@ -4811,7 +4811,7 @@ ___fast_switch_2_rx:
                     log_printf(INF, "select error.");
                 } else if ( retval != 0 ) {
                 } else { // 超时s
-                    ret == ERR_FRAME_CHECK_DATA_TOO_SHORT;
+                    ret == (int)(ERR_FRAME_CHECK_DATA_TOO_SHORT);
                     break;
                 }
                 errno = 0;
